fix(个人信息): 三方信息保存
--bug=1008394 --user=郭雨琦 【github#8173】user bind Jira id with
'not_authorized' error 用户绑定JIRA账号报未授权错误
